Refactor: Flask app to standard structure and ensured bulleted AI responses
